Add report comment button, display user ratings
