--- web/src/components/Comments.css
@@ -1,23 +1,0 @@
-.comments-container {
-  background-color: var(--beige);
-  padding: 15px;
-  border-radius: 10px;
-  margin: 2% auto;
-  width: 80%;
-  font-size: 20px;
-}
-
-
-.comment-container {
-  border: solid #a90011 2px;
-  border-radius: 8px;
-  padding: 1%;
-  font-size: 24px;
-  margin-bottom: 20px;
-}
-
-
-h2 {
-  font-size: 36px;
-  color: #a90011;
-}
