+++ web/src/components/RatingStar.css
@@ -1,14 +1,9 @@
 .rating-container {
   display: table;
+  display: flex;
 }
 
 .rating-label {
-  float: left;
-  margin-top: 3px;
   margin-left: 5px;
   color: #5a5a5a;
-}
-
-.rating-star {
-  float: left;
 }
