+++ web/src/components/product/ProductInfo.js
@@ -1,9 +1,9 @@
 import React from 'react'
 
-import WishlistButton from '../components/WishlistButton'
-import RatingStar from '../components/RatingStar'
+import WishlistButton from '../WishlistButton'
+import RatingStar from '../RatingStar'
 
-import FreeShippingImage from '../images/free-shipping.png'
+import FreeShippingImage from '../../images/free-shipping.png'
 
 import './ProductInfo.css'
